fixed snackbar in wallet
